// BSAF-SPA/src/app/models/BeneficiarySearchForm.ts
export class BeneficiarySearchForm{
    cardID: string;
    beneficiaryName?: string;
    beneficiaryFName?: string;
    borderPoint?: string;
    beneficiaryType?: string;
    returnStatus?: string;
    ScreeningDateFrom?: string;
    ScreeningDateTo?: string;
    pageIndex: number;
    length: number;
    pageSize: number;
}